Modification dossier model
