+++ src/models/Game.js
@@ -1,0 +1,11 @@
+import ObjUser from './User';
+
+export default class Game {
+    constructor () {
+        this.Users = [];
+    }
+
+    addPlayer(Username) {
+        this.Users.push(new ObjUser(Username));
+    }
+}
