+++ src/models/Score.js
@@ -1,15 +1,4 @@
 import User from './User';
-import ObjUser from './User';
-
-export default class Game {
-    constructor () {
-        this.Users = [];
-    }
-
-    addPlayer(Username) {
-        this.Users.push(new ObjUser(Username));
-    }
-}
 
 export default class Score {
     constructor () {
@@ -69,7 +58,7 @@ export default class Score {
 
     tourSuivant() {
         this.tourCourant++;
-        for (Users in User) {
+        for (this.Users in User) {
             User.tourCourant = this.tourCourant;
         }
     }
